requirements.txt
accelerate==1.6.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
diffusers==0.33.1
faiss-cpu==1.10.0
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
fsspec==2025.3.2
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pillow==11.1.0
psutil==7.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.2
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
zipp==3.21.0
